Created laser enemy
Created an enemy - as well as laser functionality for more types of laser enemies - that will remain stationary and attempt to snipe the player with a raycast laser. Laser collides with walls (in theory) and currently deletes the player on contact (may need more testing). An example of them in action is available within my scene (Theo). Will need polishing later.