Enemy attack improvement
Created a hitbox script for enemies that is highly customisable (will document/improve ease of use at a later date). The hitbox would function if the player script actually had health, but it does not currently, and I don't want to mess with their scripts in the event of a codebase resonance cascade. Also dipped toes into improving the movement of the grunts, now have an attack cooldown and lunge but getting stuck, floatiness etc not fixed yet