Wave spawn system V2.0
The W.S.S has been reconfigured into spawning a horde of random enemies in a scenes / room using triggers, unlike a set number per waves in the previous version.

A random number of total enemies to spawn will be generated, and then the system will spawn the random enemies based on how many enemies are currently on screen and the total to spawn, to prevent lag.

- The V2 W.S.S affects the script and the prefab, in order to accommodate the changes needed.

- Current issue -

- Need to find a way to have enemies deaths trigger a key function within the script so that the counter can go down and spawn a new enemy on screen.