Enemy class audio source bug
For an unknown reason, if the enemy class has the audio source disabled or removed, it will also prevent them from moving at all.

It's possible that a null or other type of error will cause the class script to stop it from working?

This push will add or enable the audio source for all affected enemy classes, as a quick fix.