Death animation fixes
Before these changes, the enemy class would enter the dead state and be destroyed immediately.

But now the dead state will enter an IEnumerator to check the current animation and it's length to ensure that the death animation for each enemy type is played and once finished, the enemy will then be destroyed immediately.