Several bug fixes and linking to get everything playable together
In this play testing session
We have enemies that spawn in rooms
doors that open and close when levels are complete
2 new mini game rooms
- colour changing floor
- safe area
new player tweaks
a game loop for testing

improvements based on last weeks playtesting feedback.

No where near as much content I hoped to add for the playtesting session(Such as boss's).