Weapons v2.4
~damage values: HG=7, MG=5, SG=3, SR=30
the intent of this damage is for the weapons to perform better against the healthier enemies. The sniper will one-shot any light foe, two shot and slightly tankier enemy and do solid damage to bosses. The shotgun is also capable of doing massive damage, but needed to keep up with the SR's massive damage so increasing the damage by one puts the SG at 45 damage if all pellets connect with a target. The pistol has the lowest damage potential out of all weapons, but each hit feels somewhat effective to compensate. The machine gun has a low damage per hit, like the shotgun, but is extremely effective at clearing out multiple enemies that are close together.
+added reload state system: reloads now automatically resume upon switching back to weapon
-removed reload button feature: reloads are triggered automatically upon reaching 0
~fixed potential bullet collision issue where bullets bounce off an immortal foe (an enemy with negative health that fails to be despawned)
~resolved shotgun inconsistent recoil direction: recoil no longer tied to instance of bullet, moved to on triggering of script
+weapon fire pitch changes as ammo loaded decreases to that end, check the next change:
~updated SFX script to play the clip to completion, including functionality to disable or enable this feature, disabled by default
+changed machine gun noise to something easier to listen to thanks to the new play to completion element
+added reload element to player hud to indicate reloading is taking place
+altered structure of reloading coroutine to make the reload finish at a consistent time.