Anti camp system V1.0
The Anti camp system is designed to prevent players from sticking to room corners for too long.

When a player enters the trigger boxes (Trigger points), it will begin a countdown to spawn turrets in the same positions as the trigger points.

However if the player enters or leaves the trigger points the countdown will reset.

Upon death of all enemies or players the Anti camp system will stop working for that room.