AntiCamp V1.1 & SpawnLogic V1.4
The AntiCamp script can now detect when to correctly shut down once all instantiated enemies from the SpawnLogic are all killed by the players.

The updated SpawnLogic will now only spawn enemies in sets of 5, and for the SpawnLogic to instantiate more enemies, the players must kill the 5 enemies on screen for another 5 to appear.