AntiCamp V1.2
The previous versions were over complicated and broke itself and the spawn Logic for no reason.

This version however has been reworked and tested in "Enemy Test" scene to act like a mine in each room.

Place the prefab in each conner of a room and if a player stays in the conner for 5 seconds, a random turret will spawn, and the Anti camp prefab will delete itself.

(Do not use SpawnLogic from this branch)